initial completion of Data Page
