+++ MarsRoverBestGroup3.0/MarsRoverBestGroup3.0/Models/Dates.cs
@@ -8,15 +8,15 @@ namespace MarsRoverBestGroup3._0.Models
     public class Dates
     {
         public DateTime earthInputDate { get; set; }
-        
-        //public double marsInputDate { get; set; }
-        
+
+        public double marsInputDate { get; set; }
+
 
     }
-    
+
     public class HomepageViewModel
     {
         public double marsOutputDate { get; set; }
-        //public DateTime earthOutputDate { get; set; }
+        public DateTime earthOutputDate { get; set; }
     }
 }
