+++ MarsRoverBestGroup3.0/MarsRoverBestGroup3.0/Views/Home/MarsData.cshtml
@@ -1,0 +1,29 @@
+﻿
+<head>
+    <link rel="stylesheet" href="/css/MarsData.css">
+</head>
+<body>
+    <section class="s_container">
+        <div class="d_header">
+            <h1>
+                Different Mars data Stuff
+            </h1>
+        </div>
+        <section class="s_container_inner">
+            <div>
+                <form method="post" action="/Home/ConvertEarthDate" class="d-flex">
+                    <input class="inputBox" id="earth-date-input" type="date" name="earthInputDate" placeholder="Earth date" aria-label="Earth date" style="text-align: center; margin-top: 10px">
+                    <button class="btn btn-outline-danger" id="earth-to-mars-convert" type="submit">Convert to</button>
+                    <output class="form-control me-2" id="mars-date-output" for="earth-date-input" aria-label="Mars Sol date" style="text-align: center; margin-top: 10px ">@Model.marsOutputDate</output>
+                </form>
+            </div>
+            <div >
+                <form method="post" action="/Home/ConvertMarsDate" class="d-flex">
+                    <input class="inputBox" id="mars-date-input" type="text" name="marsInputDate" placeholder="Mars Sol Date" aria-label="Mars Sol Date" style="text-align: center; margin-top: 10px">
+                    <button class="btn btn-outline-danger " id="mars-to-earth-convert" type="submit">Convert to</button>
+                    <output class="form-control me-2" id="earth-date-output" for="mars-date-input" aria-label="Earth date" style="text-align: center; margin-top: 10px ">@Model.earthOutputDate.ToShortDateString() (give or take a day)</output>
+                </form>
+            </div>
+        </section>
+    </section>
+</body>
