+++ MarsRoverBestGroup3.0/MarsRoverBestGroup3.0/Views/Shared/_Layout.cshtml
@@ -19,6 +19,7 @@
                     <ul>
                         <li>@Html.ActionLink("Home Page", "Index", "Home")</li>
                         <li>@Html.ActionLink("Gallery Page", "Gallery", "Home")</li>
+                        <li>@Html.ActionLink("Mars Data Page", "MarsData", "Home")</li>
                     </ul>
                 </div>
             </div>
@@ -27,7 +28,7 @@
             @RenderBody()
         </main>
         <footer>
-            <p>Mars Rover hub page designed by Claire, Sofie, Ray & Joseph :)</p>
+            <p>Mars Rover hub page designed by Communism :)</p>
             <div id="footer-nav">
                 |<a> index </a>|
                 <a> gallery </a>|
